Update the kim_property_modify
Update the kim_property_modify to the current properties framework
document. This update includes clean up the code and change the
bahvior when having standard keys with extent.
Now it is possible to support caese:
If the source-value key is a scalar, the values of the uncertainty
and digits keys must be scalars. If the source-value key’s value
is an array (EDN vector), the values of the uncertainty and digits
keys must be either arrays of the same extent, or scalars in which
case they are taken to apply equally to all values in the
source-value array.